fix(infra): reread device pairing state inside the write transaction
Every device pairing persist replaced the pending and paired tables with an
in-memory snapshot loaded before the write transaction began. The pairing lock
is process-local, and the Gateway and CLI share openclaw.sqlite, so a persist
from one process could delete rows another process had just committed: a token
last-used write from a snapshot predating a node host approval silently
unpaired that node host.

Pairing mutations now run through mutateDevicePairingStoreStateInTransaction,
which reads the authoritative snapshot inside the immediate transaction,
applies the mutation, and rewrites the tables in the same commit, matching the
channel pairing store. Row-scoped node surface updates keep the pairing lock
but drop their redundant full-snapshot load. Splits row mapping into
device-pairing-store-rows.ts and the silent-pairing prune into
device-pairing-prune.ts to keep both owners under the file size cap.
